requirements.txt
matplotlib
numpy
pandas
Pillow
scipy
seaborn
tensorboardX
torch==1.4.0
torchvision==0.5.0
tqdm
